fix(ci): decompose IssueDocument + actions.go, repair vet test callers
Lint (file-size, 1500-LOC cap):
- IssueDocument.tsx 2276 → 1425 LOC. Extracted IssueDetailTabs,
  IssueDescription, SubIssuesList, OwnerPicker, ParentIssueBreadcrumb,
  ReopenIssueButton, IssueActionToolbar into co-located files.
  ApproveAndStartButton, CloseIssueButton, CommentsTimeline, IssueComment
  exported from IssueDocument for the extracted modules.
- internal/teammcp/actions.go 1802 → 1320 LOC. Extracted
  recordExecuteAudit / sanitizeOutcomeError / outcomeDedupe* /
  brokerPostActionOutcomeMessage / brokerPostApprovalAudit into
  actions_audit.go; summarizeActionResult / shapeKnownResult /
  shapeArraySummary / shapeArrayItem into actions_result_summary.go;
  resolveActionIssue / looksLikeOpaqueID / hasLetterAndDigit /
  consonantRun into actions_resolve_issue.go.

Vet:
- Updated stale 2-arg selfHealingTaskTitle and 4-arg
  selfHealingTaskDetails call sites across 5 test files to match the
  new (agentSlug, taskID, parentTitle, reason[, detail]) signatures.

gofmt: applied to touched files.
