+++ internal/teammcp/actions.go
@@ -8,7 +8,6 @@ import (
 	"net/url"
 	"os"
 	"strings"
-	"sync"
 	"time"
 
 	"github.com/modelcontextprotocol/go-sdk/mcp"
@@ -292,145 +291,6 @@ func requireTeamActionApproval(ctx context.Context, slug, channel string, args T
 	}
 }
 
-// resolveActionIssue returns the Issue (team_task) id that scopes this
-// action call. The resolution order:
-//
-//  1. If args.IssueID is set and exists in the broker, use it.
-//  2. Otherwise, look at open Issues in this channel. If at least one
-//     exists, return the most-recently-updated one. The action gets
-//     attached to that Issue (the agent is presumed to be continuing
-//     work on the active Issue).
-//  3. Otherwise, auto-create a new draft Issue with a title derived
-//     from the action verb + platform, and return its id.
-//
-// The product rule the user locked: "any work getting done has an Issue
-// behind it." Rather than rejecting calls without issue_id and forcing
-// the LLM to self-correct, the broker resolves the container itself.
-// The agent never sees an error; the human always sees an Issue
-// scoping the work.
-func resolveActionIssue(ctx context.Context, slug, channel string, args TeamActionExecuteArgs) (string, string, error) {
-	// Step 1: caller-supplied id (still preferred — agent knows the
-	// scoping better than we do). Verify it exists; if not, fall through
-	// to the find-or-create path rather than fail. Returns the resolved
-	// id AND the Issue's current lifecycle_state so the caller can
-	// enforce gates (e.g. block external actions while the Issue is
-	// awaiting human approval in drafting).
-	if id := strings.TrimSpace(args.IssueID); id != "" {
-		var resp struct {
-			Task *struct {
-				ID             string `json:"id"`
-				LifecycleState string `json:"lifecycle_state"`
-			} `json:"task,omitempty"`
-			LifecycleState string `json:"lifecycleState"`
-		}
-		if err := brokerGetJSON(ctx, "/tasks/"+url.PathEscape(id), &resp); err == nil &&
-			resp.Task != nil && strings.TrimSpace(resp.Task.ID) != "" {
-			state := strings.TrimSpace(resp.Task.LifecycleState)
-			if state == "" {
-				state = strings.TrimSpace(resp.LifecycleState)
-			}
-			return id, state, nil
-		}
-	}
-
-	// Step 2: find the most recently updated open Issue in this channel.
-	// "Open" = anything that isn't done / approved / rejected /
-	// cancelled. Agents typically work on the latest active Issue, so
-	// preferring the newest-updated is a good default.
-	if ch := strings.TrimSpace(channel); ch != "" {
-		var list struct {
-			Tasks []struct {
-				ID         string `json:"id"`
-				Status     string `json:"status"`
-				Lifecycle  string `json:"lifecycle_state"`
-				UpdatedAt  string `json:"updated_at"`
-				CreatedAt  string `json:"created_at"`
-				ParentID   string `json:"parent_issue_id,omitempty"`
-				TaskTypeID string `json:"task_type,omitempty"`
-			} `json:"tasks"`
-		}
-		path := "/tasks?viewer_slug=" + url.QueryEscape(slug) + "&channel=" + url.QueryEscape(ch)
-		if err := brokerGetJSON(ctx, path, &list); err == nil && len(list.Tasks) > 0 {
-			var best string
-			var bestState string
-			var bestTS string
-			for _, t := range list.Tasks {
-				switch strings.ToLower(strings.TrimSpace(t.Status)) {
-				case "done", "approved", "rejected", "cancelled", "canceled":
-					continue
-				}
-				switch strings.ToLower(strings.TrimSpace(t.Lifecycle)) {
-				case "approved", "rejected":
-					continue
-				}
-				// Pick the row with the largest updated_at (RFC3339
-				// strings compare correctly), falling back to
-				// created_at when updated_at is empty.
-				ts := t.UpdatedAt
-				if ts == "" {
-					ts = t.CreatedAt
-				}
-				if ts > bestTS {
-					bestTS = ts
-					best = t.ID
-					bestState = strings.TrimSpace(t.Lifecycle)
-				}
-			}
-			if best != "" {
-				return best, bestState, nil
-			}
-		}
-	}
-
-	// Step 3: auto-create a draft Issue. Title is the human-readable
-	// verb + platform from the action (the same one that goes on the
-	// approval card). Description carries enough action context for the
-	// human to recognise what work this is — without it the Issue card
-	// would just say "Run an action via Gmail" with no further detail.
-	verb := actionVerbLabel(args.Platform, args.ActionID)
-	platformLabel := platformDisplay(args.Platform)
-	title := titleCaser.String(verb) + " via " + platformLabel
-	details := strings.TrimSpace(args.Summary)
-	if details == "" {
-		details = "Auto-created by the broker to scope an external action the agent kicked off without an explicit Issue. " +
-			"Action: " + strings.TrimSpace(args.ActionID) + " via " + platformLabel + "."
-	}
-	// created_by must be a channel member, not the literal "broker"
-	// string, because the broker's POST /task-plan ACL check rejects
-	// non-member actors with 403. The calling agent is always a member
-	// of the channel they're acting in, so attribute the auto-create to
-	// them. The "auto-created by gate" intent shows up in `details` and
-	// in the absence of an agent-authored description.
-	createBody := map[string]any{
-		"channel":    strings.TrimSpace(channel),
-		"created_by": slug,
-		"tasks": []map[string]any{
-			{
-				"title":          title,
-				"assignee":       slug,
-				"details":        details,
-				"task_type":      "issue",
-				"execution_mode": "office",
-			},
-		},
-	}
-	var created struct {
-		Tasks []struct {
-			ID string `json:"id"`
-		} `json:"tasks"`
-	}
-	if err := brokerPostJSON(ctx, "/task-plan", createBody, &created); err != nil {
-		return "", "", fmt.Errorf("auto-create issue: %w", err)
-	}
-	if len(created.Tasks) == 0 || strings.TrimSpace(created.Tasks[0].ID) == "" {
-		return "", "", fmt.Errorf("auto-create issue returned no id")
-	}
-	// Auto-created Issues land in `drafting` (via MutateTask create-path
-	// applyLifecycleStateLocked) so the human still controls the gate
-	// before the agent proceeds.
-	return strings.TrimSpace(created.Tasks[0].ID), "drafting", nil
-}
-
 // actionApprovalDedupeKey collapses retries of the same external-action
 // call onto one approval request. Keyed on agent + platform + action_id +
 // connection so an in-flight retry by the agent loop folds onto the
@@ -601,67 +461,6 @@ func actionVerbLabel(platform, actionID string) string {
 		return "run an action"
 	}
 	return strings.Join(tokens, " ")
-}
-
-// looksLikeOpaqueID reports whether s smells like an internal identifier
-// rather than a human-readable action verb. Heuristics:
-//   - contains `::` (the connection-manager internal delimiter we saw in
-//     the wild on May 26 — request-6/7 titles rendered raw hashes that way)
-//   - any whitespace/separator-split token longer than 12 chars AND
-//     containing both letters and digits (hash-shaped)
-//   - any token with 4+ consecutive consonants (gibberish-shaped)
-//
-// The bar is intentionally high so we do not eat real verbs like
-// "send_email" or "list_messages". When in doubt the caller falls back to
-// "run an action" and the human can still read the action_id from the
-// "Action:" line of the context block.
-func looksLikeOpaqueID(s string) bool {
-	if strings.Contains(s, "::") {
-		return true
-	}
-	for _, tok := range strings.FieldsFunc(s, actionIDSeparator) {
-		if len(tok) >= 12 && hasLetterAndDigit(tok) {
-			return true
-		}
-		if consonantRun(tok) >= 5 {
-			return true
-		}
-	}
-	return false
-}
-
-func hasLetterAndDigit(s string) bool {
-	var hasL, hasD bool
-	for _, r := range s {
-		switch {
-		case r >= 'a' && r <= 'z', r >= 'A' && r <= 'Z':
-			hasL = true
-		case r >= '0' && r <= '9':
-			hasD = true
-		}
-		if hasL && hasD {
-			return true
-		}
-	}
-	return false
-}
-
-func consonantRun(s string) int {
-	maxRun, run := 0, 0
-	for _, r := range s {
-		isVowel := r == 'a' || r == 'e' || r == 'i' || r == 'o' || r == 'u' ||
-			r == 'A' || r == 'E' || r == 'I' || r == 'O' || r == 'U' || r == 'y' || r == 'Y'
-		isConsonant := (r >= 'a' && r <= 'z') || (r >= 'A' && r <= 'Z')
-		if isConsonant && !isVowel {
-			run++
-			if run > maxRun {
-				maxRun = run
-			}
-		} else {
-			run = 0
-		}
-	}
-	return maxRun
 }
 
 // platformDisplay turns a kebab-cased provider slug into its human form.
@@ -1036,245 +835,6 @@ func handleTeamActionExecute(ctx context.Context, _ *mcp.CallToolRequest, args T
 	return textResult(prettyObject(result)), nil, nil
 }
 
-// recordExecuteAudit writes an ApprovalAuditEntry to the broker for an
-// executed action. No-op when the approval was bypassed (read-only, dry-run,
-// unsafe) — those calls don't have a request id to correlate to.
-func recordExecuteAudit(ctx context.Context, approvalCtx approvalContext, args TeamActionExecuteArgs, actor, channel, executedAt, outcome, summary, msgID string) {
-	if strings.TrimSpace(approvalCtx.RequestID) == "" {
-		return
-	}
-	platform := strings.TrimSpace(args.Platform)
-	if platform == "" {
-		platform = "unknown"
-	}
-	actionID := strings.TrimSpace(args.ActionID)
-	if actionID == "" {
-		actionID = "unknown"
-	}
-	brokerPostApprovalAudit(ctx, team.ApprovalAuditEntry{
-		ApprovalRequestID:    approvalCtx.RequestID,
-		TaskID:               approvalCtx.IssueID,
-		Platform:             platform,
-		ActionID:             actionID,
-		ConnectionKey:        strings.TrimSpace(args.ConnectionKey),
-		RequestedAt:          approvalCtx.RequestedAt,
-		AnsweredAt:           approvalCtx.AnsweredAt,
-		ExecutedAt:           executedAt,
-		Outcome:              outcome,
-		OutcomeSummary:       summary,
-		OutcomeChatMessageID: msgID,
-		Actor:                actor,
-		Channel:              channel,
-	})
-}
-
-// summarizeActionResult renders the action result into a short,
-// chat-friendly block the human can read at a glance. The goal is
-// "what came back?" — not the full payload, not raw JSON wrapping.
-// For action results that are obviously structured (a list, a map
-// with familiar keys), we surface the top-level shape (count, ids,
-// subjects). For opaque shapes, we fall back to a pretty-printed
-// excerpt clipped to a budget so chat stays readable.
-//
-// The agent is still prompted to post a richer human_message right
-// after — that's where the interpretation lives. This preview just
-// guarantees the raw signal is visible immediately, even before the
-// agent's followup lands.
-func summarizeActionResult(result any) string {
-	if result == nil {
-		return ""
-	}
-	// Try smart shaping for common list/map shapes first.
-	if shaped := shapeKnownResult(result); shaped != "" {
-		return shaped
-	}
-	// Fall back: pretty JSON, clipped + fenced so chat renders it
-	// as a code block.
-	raw := prettyObject(result)
-	raw = strings.TrimSpace(raw)
-	if raw == "" || raw == "null" || raw == "{}" || raw == "[]" {
-		return ""
-	}
-	const maxPreviewBytes = 600
-	if len(raw) > maxPreviewBytes {
-		raw = raw[:maxPreviewBytes] + "\n… (truncated; agent will summarize next)"
-	}
-	return "```\n" + raw + "\n```"
-}
-
-// shapeKnownResult tries to lift the most common result shapes into a
-// human line:
-//   - a slice → "N item(s)" + a couple of representative summaries
-//   - a map with a `threads` / `messages` / `items` / `results` /
-//     `events` slice → same as above
-//
-// Returns "" when no familiar shape matches; the caller falls back to
-// the raw JSON excerpt.
-func shapeKnownResult(result any) string {
-	// Slice at the top level.
-	if arr, ok := result.([]any); ok {
-		return shapeArraySummary(arr)
-	}
-	// Map at the top level — peek for a familiar list field.
-	if m, ok := result.(map[string]any); ok {
-		for _, key := range []string{"threads", "messages", "items", "results", "records", "events", "data", "rows"} {
-			if v, present := m[key]; present {
-				if arr, ok := v.([]any); ok {
-					summary := shapeArraySummary(arr)
-					if summary != "" {
-						return key + ": " + summary
-					}
-				}
-			}
-		}
-		// Single-item map with an id-shaped field is also useful
-		// (e.g. send_email returns {"id": "...", "thread_id": "..."}).
-		var ids []string
-		for _, key := range []string{"id", "thread_id", "message_id", "event_id"} {
-			if v, ok := m[key].(string); ok && v != "" {
-				ids = append(ids, key+"="+v)
-			}
-		}
-		if len(ids) > 0 {
-			return strings.Join(ids, ", ")
-		}
-	}
-	return ""
-}
-
-// shapeArraySummary builds a "N items" header + up to 3 representative
-// one-line excerpts, picking common identifying fields (subject, name,
-// title, id) when each item is a map.
-func shapeArraySummary(arr []any) string {
-	if len(arr) == 0 {
-		return "0 results"
-	}
-	var b strings.Builder
-	fmt.Fprintf(&b, "%d result(s)", len(arr))
-	max := len(arr)
-	if max > 3 {
-		max = 3
-	}
-	for i := 0; i < max; i++ {
-		line := shapeArrayItem(arr[i])
-		if line == "" {
-			continue
-		}
-		fmt.Fprintf(&b, "\n• %s", line)
-	}
-	if len(arr) > max {
-		fmt.Fprintf(&b, "\n… and %d more", len(arr)-max)
-	}
-	return b.String()
-}
-
-func shapeArrayItem(item any) string {
-	m, ok := item.(map[string]any)
-	if !ok {
-		s := strings.TrimSpace(fmt.Sprint(item))
-		if len(s) > 120 {
-			s = s[:120] + "…"
-		}
-		return s
-	}
-	// Try common identifying fields, in order of "most readable".
-	for _, key := range []string{"subject", "title", "name", "summary", "snippet", "preview"} {
-		if v, ok := m[key].(string); ok {
-			v = strings.TrimSpace(v)
-			if v != "" {
-				if len(v) > 100 {
-					v = v[:100] + "…"
-				}
-				// Tack on a from/sender when present for emails.
-				for _, fromKey := range []string{"from", "sender", "author"} {
-					if f, ok := m[fromKey].(string); ok && f != "" {
-						return v + " — " + f
-					}
-				}
-				return v
-			}
-		}
-	}
-	// Fall back to id when nothing readable found.
-	for _, key := range []string{"id", "thread_id", "message_id"} {
-		if v, ok := m[key].(string); ok && v != "" {
-			return key + "=" + v
-		}
-	}
-	return ""
-}
-
-// sanitizeOutcomeError strips structured noise (JSON event blobs, CLI
-// flow envelopes, multi-line stack traces) out of an error message so
-// the chat outcome reads as a human-facing failure reason. Without
-// this, an error like:
-//
-//	"one CLI failed: {\"event\":\"flow:start\",\"flowKey\":\"wuphf-auto-action\",...}"
-//
-// dominates the channel with implementation noise. We keep only the
-// human prefix and replace the JSON tail with `(see logs)`.
-func sanitizeOutcomeError(s string) string {
-	s = strings.TrimSpace(s)
-	if s == "" {
-		return "no detail"
-	}
-	// Strip everything from the first { or [ onward — that's where the
-	// JSON/CLI envelope starts. Keep the human prefix.
-	if i := strings.IndexAny(s, "{["); i >= 0 {
-		prefix := strings.TrimRight(strings.TrimSpace(s[:i]), ": ")
-		if prefix != "" {
-			s = prefix + " (see logs)"
-		} else {
-			s = "(see logs)"
-		}
-	}
-	if i := strings.IndexAny(s, "\r\n"); i >= 0 {
-		s = s[:i]
-	}
-	if len(s) > 200 {
-		s = s[:200] + "…"
-	}
-	return s
-}
-
-// outcomeDedupeWindow is how long an identical action outcome counts
-// as "just posted" — within this window, repeated successful execs of
-// the same action by the same agent in the same channel are silenced
-// so the channel doesn't fan out into "✅ ✅ ✅" spam. Failures are
-// never deduped (silence-on-failure is the bug we never want).
-const outcomeDedupeWindow = 8 * time.Second
-
-var (
-	outcomeDedupeMu  sync.Mutex
-	outcomeDedupeMap = map[string]time.Time{}
-)
-
-// recentlyPostedOutcome returns true when the same (slug, action,
-// channel) tuple has been posted in the last outcomeDedupeWindow.
-// Also stamps the current time on the key, so callers don't need to
-// record-then-check separately.
-func recentlyPostedOutcome(slug, actionID, channel string) bool {
-	key := strings.ToLower(strings.TrimSpace(slug)) + "|" +
-		strings.ToLower(strings.TrimSpace(actionID)) + "|" +
-		strings.ToLower(strings.TrimSpace(channel))
-	now := time.Now()
-	outcomeDedupeMu.Lock()
-	defer outcomeDedupeMu.Unlock()
-	if last, ok := outcomeDedupeMap[key]; ok && now.Sub(last) < outcomeDedupeWindow {
-		return true
-	}
-	outcomeDedupeMap[key] = now
-	// Opportunistic GC: prune entries older than 10x the window so the
-	// map doesn't grow unbounded across long sessions.
-	cutoff := now.Add(-outcomeDedupeWindow * 10)
-	for k, t := range outcomeDedupeMap {
-		if t.Before(cutoff) {
-			delete(outcomeDedupeMap, k)
-		}
-	}
-	return false
-}
-
 func handleTeamActionWorkflowCreate(ctx context.Context, _ *mcp.CallToolRequest, args TeamActionWorkflowCreateArgs) (*mcp.CallToolResult, any, error) {
 	slug, err := resolveSlug(args.MySlug)
 	if err != nil {
@@ -1568,57 +1128,6 @@ func brokerRecordAction(ctx context.Context, kind, source, channel, actor, summa
 		"summary":    strings.TrimSpace(summary),
 		"related_id": strings.TrimSpace(relatedID),
 	}, nil)
-}
-
-// brokerPostActionOutcomeMessage posts a system-authored chat message
-// summarising the outcome of an external action that just ran. Without
-// this, the human's only signal that an approved action actually
-// executed was the agent's *next* message — and if the agent went
-// silent (timed out, lost context, didn't follow the prompt rule to
-// surface outcomes), the human saw nothing. This guarantees a visible
-// trace in the channel whenever an action runs, success or failure.
-//
-// Channel posting goes through the broker as actor=ceo so it routes
-// through normal channel auth + the message ends up in the chat
-// surface the human is already watching. `system` would be cleaner
-// semantically but isn't a registered agent slug — the broker rejects
-// posts from unknown senders, so we use the lead.
-func brokerPostActionOutcomeMessage(ctx context.Context, channel, actor, content string) string {
-	if strings.TrimSpace(content) == "" {
-		return ""
-	}
-	ch := resolveChannel(channel)
-	from := strings.TrimSpace(actor)
-	if from == "" {
-		from = "ceo"
-	}
-	var resp struct {
-		ID string `json:"id"`
-	}
-	if err := brokerPostJSON(ctx, "/messages", map[string]any{
-		"from":    from,
-		"channel": ch,
-		"kind":    "action_outcome",
-		"content": strings.TrimSpace(content),
-	}, &resp); err != nil {
-		// Posting outcome messages is best-effort; failures here would
-		// drop the chat trace but not block execution. Returning an empty
-		// id surfaces to the audit-entry writer as "no deep link" rather
-		// than blocking the audit record entirely.
-		return ""
-	}
-	return strings.TrimSpace(resp.ID)
-}
-
-// brokerPostApprovalAudit ships an audit entry to the broker. Best-effort:
-// errors are swallowed because audit failures must never block the action
-// result from returning to the agent (which is already either committed or
-// rolled back at this point).
-func brokerPostApprovalAudit(ctx context.Context, entry team.ApprovalAuditEntry) {
-	if strings.TrimSpace(entry.ApprovalRequestID) == "" {
-		return
-	}
-	_ = brokerPostJSON(ctx, "/approval-audit", entry, nil)
 }
 
 type workflowSkillSpec struct {
